requirements.txt
pandas
scikit-learn
SQLAlchemy
apache-airflow
mlflow
boto3
marker-pdf